+++ .github/workflows/assign-manually.yml
@@ -8,8 +8,8 @@ permissions:
   issues: read
 
 jobs:
-  assign-status-sync:
-    uses: ./.github/workflows/assign-status-sync.yml
+  status-update-call:
+    uses: ./.github/workflows/status-update-call.yml
     with:
       issue_number: ${{ github.event.issue.number }}
       action: ${{ github.event.action }}
